Re-arrange completion modules
