+++ unison-cli/src/Unison/CommandLine.hs
@@ -17,18 +17,6 @@ module Unison.CommandLine
     warn,
     warnNote,
 
-    -- * Completers
-    completion,
-    completion',
-    exactComplete,
-    fuzzyComplete,
-    fuzzyCompleteHashQualified,
-    prefixIncomplete,
-    prettyCompletion,
-    prettyCompletionWithQueryPrefix,
-    fixupCompletion,
-    completeWithinQueryNamespace,
-
     -- * Other
     parseInput,
     prompt,
@@ -44,12 +32,9 @@ import Data.Configurator (autoConfig, autoReload)
 import qualified Data.Configurator as Config
 import Data.Configurator.Types (Config, Worth (..))
 import Data.List (isPrefixOf, isSuffixOf)
-import qualified Data.List as List
-import Data.List.Extra (nubOrd)
 import Data.ListLike (ListLike)
 import qualified Data.Map as Map
 import qualified Data.Text as Text
-import qualified System.Console.Haskeline as Line
 import System.FilePath (takeFileName)
 import Text.Regex.TDFA ((=~))
 import Unison.Codebase.Branch (Branch0)
@@ -59,13 +44,8 @@ import qualified Unison.Codebase.Watch as Watch
 import qualified Unison.CommandLine.Globbing as Globbing
 import Unison.CommandLine.InputPattern (InputPattern (..))
 import qualified Unison.CommandLine.InputPattern as InputPattern
-import qualified Unison.HashQualified as HQ
-import qualified Unison.HashQualified' as HQ'
-import Unison.Names (Names)
 import Unison.Prelude
-import qualified Unison.Server.SearchResult as SR
 import qualified Unison.Util.ColorText as CT
-import qualified Unison.Util.Find as Find
 import qualified Unison.Util.Pretty as P
 import Unison.Util.TQueue (TQueue)
 import qualified Unison.Util.TQueue as Q
@@ -125,95 +105,6 @@ emojiNote lead s = P.group (fromString lead) <> "\n" <> P.wrap s
 
 nothingTodo :: (ListLike s Char, IsString s) => P.Pretty s -> P.Pretty s
 nothingTodo = emojiNote "😶"
-
-completion :: String -> Line.Completion
-completion s = Line.Completion s s True
-
-completion' :: String -> Line.Completion
-completion' s = Line.Completion s s False
-
--- discards formatting in favor of better alignment
--- prettyCompletion (s, p) = Line.Completion s (P.toPlainUnbroken p) True
--- preserves formatting, but Haskeline doesn't know how to align
-prettyCompletion :: Bool -> (String, P.Pretty P.ColorText) -> Line.Completion
-prettyCompletion endWithSpace (s, p) = Line.Completion s (P.toAnsiUnbroken p) endWithSpace
-
--- | Renders a completion option with the prefix matching the query greyed out.
-prettyCompletionWithQueryPrefix ::
-  Bool ->
-  -- | query
-  String ->
-  -- | completion
-  String ->
-  Line.Completion
-prettyCompletionWithQueryPrefix endWithSpace query s =
-  let coloredMatch = P.hiBlack (P.string query) <> P.string (drop (length query) s)
-   in Line.Completion s (P.toAnsiUnbroken coloredMatch) endWithSpace
-
-fuzzyCompleteHashQualified :: Names -> String -> [Line.Completion]
-fuzzyCompleteHashQualified b q0@(HQ'.fromString -> query) = case query of
-  Nothing -> []
-  Just query ->
-    fixupCompletion q0 $
-      makeCompletion <$> Find.fuzzyFindInBranch b query
-  where
-    makeCompletion (sr, p) =
-      prettyCompletion False (HQ.toString . SR.name $ sr, p)
-
-fuzzyComplete :: String -> [String] -> [Line.Completion]
-fuzzyComplete absQuery@('.' : _) ss = completeWithinQueryNamespace absQuery ss
-fuzzyComplete fuzzyQuery ss =
-  fixupCompletion fuzzyQuery (prettyCompletion False <$> Find.simpleFuzzyFinder fuzzyQuery ss id)
-
--- | Constructs a list of 'Completion's from a query and completion options by
--- filtering them for prefix matches. A completion will be selected if it's an exact match for
--- a provided option.
-exactComplete :: String -> [String] -> [Line.Completion]
-exactComplete q ss = go <$> filter (isPrefixOf q) ss
-  where
-    go s = prettyCompletionWithQueryPrefix (s == q) q s
-
--- | Completes a list of options, limiting options to the same namespace as the query,
--- or the namespace's children if the query is itself a namespace.
---
--- E.g.
--- query: "base"
--- would match: ["base", "base.List", "base2"]
--- wouldn't match: ["base.List.map", "contrib", "base2.List"]
-completeWithinQueryNamespace :: String -> [String] -> [Line.Completion]
-completeWithinQueryNamespace q ss = (go <$> (limitToQueryNamespace q $ ss))
-  where
-    go s = prettyCompletionWithQueryPrefix (s == q) q s
-    limitToQueryNamespace :: String -> [String] -> [String]
-    limitToQueryNamespace query xs =
-      nubOrd $ catMaybes (fmap ((query <>) . thing) . List.stripPrefix query <$> xs)
-      where
-        thing ('.' : rest) = '.' : takeWhile (/= '.') rest
-        thing other = takeWhile (/= '.') other
-
-prefixIncomplete :: String -> [String] -> [Line.Completion]
-prefixIncomplete q ss = go <$> filter (isPrefixOf q) ss
-  where
-    go s =
-      prettyCompletion
-        False
-        (s, P.hiBlack (P.string q) <> P.string (drop (length q) s))
-
--- workaround for https://github.com/judah/haskeline/issues/100
--- if the common prefix of all the completions is smaller than
--- the query, we make all the replacements equal to the query,
--- which will preserve what the user has typed
-fixupCompletion :: String -> [Line.Completion] -> [Line.Completion]
-fixupCompletion _q [] = []
-fixupCompletion _q [c] = [c]
-fixupCompletion q cs@(h : t) =
-  let commonPrefix (h1 : t1) (h2 : t2) | h1 == h2 = h1 : commonPrefix t1 t2
-      commonPrefix _ _ = ""
-      overallCommonPrefix =
-        foldl commonPrefix (Line.replacement h) (Line.replacement <$> t)
-   in if not (q `isPrefixOf` overallCommonPrefix)
-        then [c {Line.replacement = q} | c <- cs]
-        else cs
 
 parseInput ::
   -- | Root branch, used to expand globs
